+++ README.md
@@ -1,5 +1,20 @@
-The file you should be modifying, Lab.java, can be found in your editor's file explorer, under the folders src/main/java. All lab instructions are contained within. While you may freely modify Application.java, no changes made to Application.java will solve the lab.
+# Hello World Lab
 
-You should run the test cases, which will verify that your code was written correctly, by clicking the 'run' button in the HelloWorldTest.java file in src/test/java. The test output will determine if your solution was correct or incorrect by comparing 'expected' and 'actual' results.
+## Pre-Requisites
+- You are expected to review the `Entities of Java` written lecture prior to working on this lab. If you are not familiar with the concepts being explored in the lab, please make sure to review the study material first.
 
-You should NEVER change the code in test files, as you would be likely to break the lab structure.
+## Description
+In this exercise, you will practice working with methods and returning a value from a method.
+
+## Completing the Lab
+You will be editing the file `Greeting.java`. This file can be found under the `src/main/java` package. Review the file's code and code comments for further instructions.
+
+Tasks to complete within the `Greeting.java` file:
+- [ ] Revise the return statement of the method named `sayHello`
+
+## Running Tests
+If you look within the `src/test/java` folder, you will find the Java class `GreetingTest.java`. The `src/test/java` folder will contain your test files.  
+
+To run the tests, click on the Testing tab (test beaker icon on the left-hand side) and run the test cases by using the "Run Test" button (play icon). Please ensure that you do not already have the program running.
+
+You can optionally run the program for manual testing by switching to the file named `Main.java` under the `src/main/java` package and clicking the "Run Java" button (play icon) on the top-right of VSCode.
