--- src/main/java/Lab.java
@@ -1,20 +1,0 @@
-public class Lab {
-    /**
-     * TODO: Change this method so that it returns "Hello, world!"
-     *
-     * This method currently contains the line:
-     * return "change me";
-     * That means that this method is a piece of code that produces a String "change me", 
-     * and when this method is used in other parts of my program, the value "change me"
-     * is returned.
-     * 
-     * Instead, this method must return a hello world String, instead of returning 
-     * "change me". A test case will check if the behavior is as expected.
-     * You should use the test results to check for differences.
-     * 
-     * @return "Hello, world!"
-     */
-    public String sayHello(){
-        return "change me";
-    }
-}
